+++ NEWS.md
@@ -4,6 +4,7 @@
 * Added support for track attributes to logical tracks. 
 * Bug fix: dbs order was changed to lexicographical order on emr_db.connect.
 * Bug fix: stray logical tracks on non-global dbs were shown.
+* Bug fix: `emr_track.attr.get` returned attribute name instaed of value.
 * Make sure logical tracks cannot be created on non-global db. 
 * Added `db_id` parameter to `emr_track.ls`
 * Removed cpp11 package dependency